Fail early on UnicodeDecodeError
If at any point there's an issue decoding either a key or value when
calling the C implementation of loads, we need to cleanup our return
value and exit early. Returning NULL will signal to Python that we
should re-raise the UnicodeDecodeError.

Also let's only check the type of the return_value once and use a
boolean local access to avoid the overhead of adding to and popping the
call stack.